# Optimize Engine for multi-core CPUs