Optimize the `Engine` for multi-core CPU capabilities.

* **src/engine.rs**
  - Use `tokio::task::spawn_blocking` for `get`, `set`, `del`, and `scan` methods to leverage multi-core CPUs.

* **benches/engine_benchmark.rs**
  - Update benchmarks to use optimized multi-core `Engine` operations.
  - Correct the usage of `rand::thread_rng()` instead of `rand::rng()`.

* **examples/concurrent_usage.rs**
  - Remove redundant header comment.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/jackysp/tegdb/pull/8?shareId=2fad4166-d8de-45df-8446-94408716570c).