@jackysp 2025-02-10
Before this PR:

```text
Performance over 10s runtime:
 Number of threads: 7
 Average set() latency: 1.551µs (total calls: 45137196, calls/sec: 4513719.60)
 Average get() latency: 284ns (total calls: 246569898, calls/sec: 24656989.80)
```

After this PR:

```text
Performance over 10s runtime:
 Number of threads: 7
 Average set() latency: 23.636µs (total calls: 2961541, calls/sec: 296154.10)
 Average get() latency: 20.844µs (total calls: 3358279, calls/sec: 335827.90)
```